co-locate signer util tests
